organized the view routes
